Rework the validation algorithm section
Add the intermediates since they come out of nowhere in the current
description of the algorithm. Adjust language to use a single trust
anchor.

Singling out 'The Version, Issuer, and Subject fields' is a bit strange.
This has the effect that Serial Number, Signature Algorithm and Subject
Public Key Info appear to be missing from the list, so simplify the
text to 'The fields'.

Remove point 5 allowing for the RFC 8360 policy and require RFC 6484
(implicitly via 6487 4.8.9).

Part of #5